+++ nbsrc/__init__.py
@@ -1,17 +1,3 @@
-"""Jupyter notebooks as python or R script
+"""The nbsrc package is empty, and only serves to expose documentation.
 
-Use the 'nbsrc' conversion script to convert Jupyter notebooks from/to
-R or Python scripts
-
-NB: read, write methods, as well as ContentsManager, are to be found in the
-nbrmd package
-"""
-
-from .nbsrc import readme
-
-try:
-    from .srcexporter import PyNotebookExporter
-    from .srcexporter import RNotebookExporter
-except ImportError as error:
-    PyNotebookExporter = str(error)
-    RNotebookExporter = str(error)
+Please use nbrmd instead"""