+++ tests/test_readme.py
@@ -1,5 +1,0 @@
-from nbsrc import readme
-
-
-def test_readme():
-    readme()